Agregue servicio de SELLER y complete entidades relacionandolas.

+++ src/main/java/com/credicoop/utn/entities/Customize.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
+import java.util.List;
 
 
 @Getter
@@ -19,7 +20,7 @@ public class Customize {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(name="customize_type")
     private String customizeType;
     @Column(name="customize_price")
@@ -28,5 +29,6 @@ public class Customize {
     private String image;
     @Column(name="phrase")
     private String phrase;
+
 
 }
